requirements.txt
Pillow
streamlit
openai
sentence-transformers
python-dotenv